% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIREcaller.R
\name{FIREcaller}
\alias{FIREcaller}
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIREcaller.R
\name{FIREcaller}
\alias{FIREcaller}
\title{FIREcaller: an R package for detecting frequently interacting regions from Hi-C data}
\usage{
FIREcaller(prefix.list=(...), gb=c("hg19","GRCh38","mm9","mm10"), map_file="", rm_mhc=c("TRUE","FALSE"),rm_EBL=c("TRUE","FALSE"),upper_cis=200000, dist=c('poisson','nb'), rm_perc=0.25)
}
\arguments{
\item{prefix.list}{a list of samples that correspond to the names of the gzipped files.}

\item{gb}{a string that defines the genome build type.If missing, an error message is returned.}

\item{map_file}{a string that defines the name of the mappability file specific to the samples genome build, restriction enzyme, and resolution. Only contains chromosomes you want to input. See read me for format.}

\item{rm_mhc}{a logicalvalue indicating whether to remove the mhc region of the sample. Default is "TRUE".}

\item{rm_EBL}{a logical value indicating whether to remove the ENCODE blacklist regions of the sample. Default is "TRUE".}

\item{upper_cis}{is a bound for the cis-interactions calculation. The default is 200Kb}

\item{dist}{is the distribution specification for the HiCNormCis normalization and FIREscore calculation. The default is Poisson.}

\item{rm_perc}{is the percentage of "bad-bins" in a cis-interaction calculation to filter. Default is 25%}
}
}
\value{
Two sets of files will be returned. The total number of files outputted are 1+ (number of prefixes/samples):

\itemize{
    \item Fire: a single text file is outputted for all the samples and all chromosomes.This file contains the Fire Score, associated ln(pvalue), and an indicator if the region is a FIRE or not with I(pvalues > -ln(0.05)).
    \item SuperFire: a text file for each sample with a list of Super Fires and corresponding -log10(pvalue).
    }
}
\description{
This function FIREcaller is a user-friendly R package for detecting FIREs from Hi-C data.
FIREcaller takes raw Hi-C NxN contact matrix as input, performs within-sample and cross-sample normaliza-tion via
HiCNormCis and quantile normalization respectively, and outputs FIRE scores, FIREs and super-FIREs.
The input file needs to be ATLEAST a NxN upper-triangular matrix OR a NxN symmetric matrix
}
\details{
The process includes calculating the raw fire scores, filtering (with the option of removing the MHC region and ENCODE black list), HiCNormCis, Quantile Normalization (if number of samples > 1), highlighting significant Fire Scores, and calculating the Super Fires.
}
\note{
Each sample must have a NxN contact frequency matrix for all autosomal chromosomes.
The prefix.list corresponds to the naming convention of the NxN contact frequency matrices.
If length(prefix.list)>1, quantile normalization is performed.
}
\examples{
# set working directory: the location of the NxN matrices and mappability files
setwd('~/Documents/Schmitt_Hippo_40KB_input/')

# define the prefix.list according to the naming convention of the NxN matrices
prefix.list <- c('Hippo')

#define the genome build
gb<-'hg19'

#define the name of the mappability file
map_file<-'F_GC_M_HindIII_40KB_hg19.txt.gz'

#define whether to remove MHC region; default=TRUE
rm_mhc <- TRUE

#' #define whether to remove ENCODE blacklist region; default=TRUE
rm_EBL<- TRUE

# define the upper bound of the cis-interactions; default=200,000; if not a multiple of the bin, then takes the ceiling;
upper_cis<-200000

# define whether the distribution should be poisson or negative binomial; default=poisson
dist<-'poisson'

# define the percentage to remove problematic bins (0-1)
rm_perc<-0.25

#run the function
FIREcaller(prefix.list,gb, map_file, rm_mhc, rm_EBL, upper_cis, dist,rm_perc)
}
\references{
Cheynna Crowley, Yuchen Yang, Ming Hu, Yun Li. FIREcaller: Detecting Frequently Interacting Regions from Hi-C Data
}
\seealso{
Paper: https://doi.org/10.1016/j.celrep.2016.10.061
}
\author{
Crowley, Cheynna Anne <cacrowle@live.unc.edu>, Yuchen Yang <yyuchen@email.unc.edu>,
Ming Hu <afhuming@gmail.com>, Yun Li <yunli@med.unc.edu>
}
