% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_interactions.R
\name{InteractionCompare}
\alias{InteractionCompare}
\title{FIREcaller: an R package for detecting frequently interacting regions from Hi-C data}
\usage{
InteractionCompare(
  data,
  targetFile,
  coef = 2,
  adjust.method = "BH",
  adj.pvalue = 0.05,
  lfc = log2(2)
)
}
\arguments{
\item{data}{a data frame of normalized cis-interactions, which are computed by FIREcaller, for each sample/replicate.}

\item{targetFile}{a vector specifying the class of each sample/replicate.}

\item{coef}{specifies the column number or column name for the interested coefficient or contrast of the linear model.}

\item{adjust.method}{specifies the method used to adjust the p-values for multiple testing. Possible options include "none", "BH", "BY" and "holm". Default is \code{adjust.method = "BH"}.}

\item{adj.pvalue}{sets the cutoff value for adjusted p-values where only the interactions with lower adjusted p-values will be returned in the output.}

\item{lfc}{sets the minimum absolute log2-fold-change required. Only the interactions with larger absolute log2-fold-change value are returned in the output.}
}
\description{
Seperate function for calling differential FIREs outside of the main FIREcaller function
}
